[Feat] main 파일분리
